backend/requirements.txt
fastapi==0.135.2
uvicorn==0.42.0
pydantic==2.12.5
python-multipart==0.0.22
openai==2.30.0
sqlalchemy==2.0.48
pypdf==6.9.2
python-docx==1.2.0
